refactor(tests): extract hash_embedding.rs with module-level cfg
Move 5 HashEmbeddingProvider tests from engine/tests/mod.rs into
engine/tests/hash_embedding.rs. Per-test #[cfg(feature = "semantic-search")]
becomes a single module-level gate on the `mod hash_embedding`
declaration in mod.rs -- matches the existing concurrency/preview_profiling
pattern.

Tests moved:
  fnv1a32_is_stable_and_deterministic
  hash_embedding_output_is_normalized_and_correct_dims
  hash_embedding_is_deterministic
  hash_embedding_rejects_empty_text
  hash_embedding_rejects_zero_dims

Private `fnv1a32` from engine/mod.rs is reachable via `use super::*;`
(descendant-private visibility).

Tests: 135 passed / 0 failed / 1 ignored (baseline unchanged under
--no-default-features; hash_embedding tests compile out as before).

marky-n1h step 4.
